feat(User): Create code to guarantee use default user picture
